// src/interfaces/interfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
  correo: string;
  password: string;
}

export interface RegisterData {
  nombre?: string;
  correo: string;
  password: string;
}

export interface LoginResponse {
  user: User;
  token: string;
}

export interface User {
  nombre: string;
  correo: string;
  rol: string;
  estado: boolean;
  google: boolean;
  uid: string;
  img?: string;
}

//Products
// Generated by https://quicktype.io

export interface ProductsResponse {
  total: number;
  productos: Producto[];
}

export interface Producto {
  _id: string;
  nombre: string;
  usuario: Categoria;
  precio: number;
  categoria: Categoria;
  disponible: boolean;
  img?: string;
}

// export interface Categoria {
//   _id: string;
//   nombre: string;
// }

//Categorias
// Generated by https://quicktype.io

export interface CategoriesResponse {
  total: number;
  categorias: Categoria[];
}

export interface Categoria {
  _id: string;
  nombre: string;
  usuario: CreadoPor;
}

export interface CreadoPor {
  _id: string;
  nombre: string;
}